chore(*): moving the server files under projects/
