+++ projects/bellumgens/server-bellumgens.ts
@@ -12,9 +12,9 @@ import * as express from 'express';
 import * as xmlhttprequest from 'xmlhttprequest';
 import { join } from 'path';
 
-import { AppServerModule } from './projects/bellumgens/src/main.server';
+import { AppServerModule } from './src/main.server';
 import { existsSync } from 'fs';
-import { environment } from './projects/common/src/environments/environment';
+import { environment } from '../common/src/environments/environment';
 
 // HTML polyfills
 (global as any).XMLHttpRequest = xmlhttprequest.XMLHttpRequest;
@@ -86,4 +86,4 @@ if (moduleFilename === __filename || moduleFilename.includes('iisnode')) {
   run();
 }
 
-export * from './projects/bellumgens/src/main.server';
+export * from './src/main.server';
