+++ projects/bellumgens/tsconfig.server.json
@@ -9,6 +9,6 @@
   },
   "files": [
     "src/main.server.ts",
-    "../../server-bellumgens.ts"
+    "server-bellumgens.ts"
   ]
 }
